# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApi",
    "version": "v1"
  },
  "paths": {
    "/api/Positions/GetAll": {
      "get": {
        "tags": [
          "Positions"
        ],
        "summary": "Obtiene todas las posiciones existentes",
        "operationId": "GetAllPositions",
        "responses": {
          "200": {
            "description": "Lista obtenida correctamente",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta API Key",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "API Key inválida",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Positions/Get/{id}": {
      "get": {
        "tags": [
          "Positions"
        ],
        "summary": "Obtiene la posición a la que pertenece el id que viene por parámetro",
        "operationId": "GetPositionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la posición",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posición encontrada",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              }
            }
          },
          "404": {
            "description": "No existe la posición",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Falta API Key",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "API Key inválida",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Positions/Create": {
      "post": {
        "tags": [
          "Positions"
        ],
        "summary": "Inserta una nueva posición",
        "operationId": "AddPosition",
        "requestBody": {
          "description": "Datos de la posición",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PositionDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado exitosamente",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Falta API Key",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "API Key inválida",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Positions/Update/{id}": {
      "put": {
        "tags": [
          "Positions"
        ],
        "summary": "Actualiza la posición a la cual pertenece el id que viene por parámetro",
        "operationId": "UpdatePosition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la posición",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Datos actualizados",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PositionDTO"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Actualizado correctamente"
          },
          "404": {
            "description": "No se encontró la posición",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Falta API Key",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "API Key inválida",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Positions/Delete/{id}": {
      "delete": {
        "tags": [
          "Positions"
        ],
        "summary": "Elimina la posición a la cual pertenece el id que viene por parámetro",
        "operationId": "DeletePosition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la posición",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado correctamente"
          },
          "404": {
            "description": "No se encontró la posición",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Falta API Key",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "API Key inválida",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Positions/Filter": {
      "get": {
        "tags": [
          "Positions"
        ],
        "summary": "Obtiene posiciones filtradas",
        "operationId": "GetFilteredPositions",
        "parameters": [
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista obtenida correctamente",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de búsqueda inválidos",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Falta API Key"
          },
          "403": {
            "description": "API Key inválida"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PositionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador del puesto.",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "description": "Título del puesto. Máximo 100 caracteres.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada del puesto. Máximo 1000 caracteres.",
            "nullable": true
          },
          "location": {
            "type": "string",
            "description": "Ubicación del puesto.",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Estado de la posición: draft, open, closed o archived.",
            "nullable": true
          },
          "recruiterId": {
            "type": "integer",
            "description": "Identificador del reclutador asignado.",
            "format": "int32"
          },
          "departmentId": {
            "type": "integer",
            "description": "Identificador del departamento correspondiente.",
            "format": "int32"
          },
          "budget": {
            "type": "number",
            "description": "Presupuesto asignado a la posición.",
            "format": "double"
          },
          "closingDate": {
            "type": "string",
            "description": "Fecha límite para aplicar al puesto. Opcional.",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Modelo de datos para una posición laboral."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Clave de autenticación usando encabezado 'X-Api-Key'. Ej: X-Api-Key: {clave}",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}